+++ dabhounds/__init__.py
@@ -1,2 +1,2 @@
 # dabhounds/__init__.py
-__version__ = "2.0.3"
+__version__ = "2.0.4"